+++ lib/generators/react_on_rails/templates/base/base/config/initializers/react_on_rails.rb
@@ -6,7 +6,8 @@
 ReactOnRails.configure do |config|
   # This configures the script to run to build the production assets by webpack. Set this to nil
   # if you don't want react_on_rails building this file for you.
-  config.build_production_command = "RAILS_ENV=production NODE_ENV=production bin/webpack"
+  # If nil, then the standard rails/webpacker assets:precompile will run
+  # config.build_production_command = nil
 
   ################################################################################
   ################################################################################